requirements.txt
pyrogram
tgcrypto
google-generativeai
aiohttp
